requirements.txt
fastapi[standard]
langchain
langchain_core
langchain_chroma
langchain_community
langchain_google_genai
pypdf
streamlit
python-dotenv
google-generativeai
